+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+- v0.1.3 - 2023-08-18
+    - do not set hasHeader with true if all headers are empty strings.
+    - added a new method: HasHeaders.
 - v0.1.2 - 2023-06-27
     - fix setting a global MaxWidth short than cell texts.
 - v0.1.1 - 2023-06-27